Move namespaces to data layer
