+++ src/domain/instance/usecases/ListInstancesUseCase.ts
@@ -2,7 +2,7 @@ import _ from "lodash";
 import { UseCase } from "../../common/entities/UseCase";
 import { RepositoryFactory } from "../../common/factories/RepositoryFactory";
 import { Repositories } from "../../Repositories";
-import { Namespace } from "../../storage/Namespaces";
+import { Namespace } from "../../../data/storage/Namespaces";
 import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { Instance, InstanceData } from "../entities/Instance";
