+++ src/domain/mapping/usecases/SaveMappingUseCase.ts
@@ -1,7 +1,7 @@
 import { Either } from "../../common/entities/Either";
 import { UseCase } from "../../common/entities/UseCase";
 import { Instance, InstanceData } from "../../instance/entities/Instance";
-import { Namespace } from "../../storage/Namespaces";
+import { Namespace } from "../../../data/storage/Namespaces";
 import { StorageClient } from "../../storage/repositories/StorageClient";
 import { DataSourceMapping } from "../entities/DataSourceMapping";
 import { isMappingOwnerStore } from "../entities/MappingOwner";
