+++ src/domain/metadata/usecases/SetResponsiblesUseCase.ts
@@ -6,7 +6,7 @@ import { RepositoryFactory } from "../../common/factories/RepositoryFactory";
 import { Instance } from "../../instance/entities/Instance";
 import { ReceivedPullRequestNotification } from "../../notifications/entities/PullRequestNotification";
 import { Repositories } from "../../Repositories";
-import { Namespace } from "../../storage/Namespaces";
+import { Namespace } from "../../../data/storage/Namespaces";
 import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { MetadataResponsible } from "../entities/MetadataResponsible";
