+++ src/domain/packages/usecases/PublishStorePackageUseCase.ts
@@ -6,7 +6,7 @@ import { RepositoryFactory } from "../../common/factories/RepositoryFactory";
 import { Instance } from "../../instance/entities/Instance";
 import { BaseModule } from "../../modules/entities/Module";
 import { Repositories } from "../../Repositories";
-import { Namespace } from "../../storage/Namespaces";
+import { Namespace } from "../../../data/storage/Namespaces";
 import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { GitHubError } from "../entities/Errors";
 import { BasePackage } from "../entities/Package";
