+++ src/domain/stores/usecases/GetStoreUseCase.ts
@@ -1,5 +1,5 @@
 import { UseCase } from "../../common/entities/UseCase";
-import { Namespace } from "../../storage/Namespaces";
+import { Namespace } from "../../../data/storage/Namespaces";
 import { StorageClient } from "../../storage/repositories/StorageClient";
 import { Store } from "../entities/Store";
