+++ src/domain/stores/usecases/SetStoreAsDefaultUseCase.ts
@@ -1,6 +1,6 @@
 import { Either } from "../../common/entities/Either";
 import { UseCase } from "../../common/entities/UseCase";
-import { Namespace } from "../../storage/Namespaces";
+import { Namespace } from "../../../data/storage/Namespaces";
 import { StorageClient } from "../../storage/repositories/StorageClient";
 import { Store } from "../entities/Store";
